MayaImport and TypeInfo now have only 2 configurations (Debug and Release). In case of "Debug with Inlines" or "Debug with Inlines and Memory Log", Debug configuration of them is used.
Also, normalized which projects must be built.
Now all projects are always built, except for that MayaImport is NOT built for 64-bit configurations.
Note that MayaImport is always built in 32-bit mode.
This should remind us regularly that it is broken =(
If you don't like it, simply unload this project from solution locally.