#4822. Use minizip as third-party library: only additional functions belong to TDM now.
Only minizip_extra.{c|h} are compiled as a part of TDM, which contain all the custom minizip functions.
In fact, two features are currently present there:
1) Some sort of file handle cloning.
2) Fast seeking in uncompressed files.

The downside of this extraction is we have to redefine all unzip's private data structures in minizip_extra.c.
If minizip changes their them, we will have to adapt this extra code immediately.