#5027. Implemented idSIMD_SSE2::CreateVertexProgramShadowCache.
It took significant time in "Behind Closed Doors" scene in "Debug with Inlines".
So I decided to implement it in idSIMD_SSE2, since optimization is force-enabled here for this build.

Also minor change in idDenseHash::Find, which might make a difference in non-optimized build.
But I'm afraid it does not =(