#4957. If idImage was not found and defaulted, don't reload it on future requests if settings differ.
I need this for particle collisions, because it loads its idImage every frame.
If the image file is missing, then a new idImage is created and "defaulted".
However, _default image uses linear filtering, while collision map uses nearest.
As the result, the default image is not reused on next frame, and yet another image is created instead.
This leads to console warnings spam and severe memory leaking.