#5137. Use double precision in dmap's mesh optimization algorithm.
Using double precision in all cross/dot products allows to produce result with perfect precision of 32-bit float in most cases.
Note that previously the same effect was achieved due to 80-bit internal precision of x87.

Having perfect precision is required for "discrete" surfaces, because they can still have T-junctions at this moment.
For any T-junction in input, an intersection point is generated in SplitOriginalEdgesAtCrossings.
But if this intersection point does not match any existing vertex exactly, then a new vertex is generated.
With imprecise computations, this quickly leads to many almost equal vertices, and discrete patches are not optimized as the result.